// types/course.ts
export interface Course {
  id: number;

  slug: string;

  title: string;

  description: string;

  category: string;

  level: string;

  duration: string;

  students: string;

  rating: string;

  price: string;

  image: string;

  imageColor: string;

  featured: boolean;

  curriculum: string[];
}